// src/generated/client/schemas/userInfoSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import { z } from "zod/v4";

export const userInfoSchema = z.object({
    "id": z.optional(z.string().describe("The user ID.")),
"name": z.optional(z.string().describe("The user's display name.")),
"url": z.optional(z.string().describe("The user's public profile URL."))
    })